requirements.txt
google-cloud-vision==2.6.3
google-cloud-storage==1.44.0
colorthief==0.2.1
scipy==1.7.3
webcolors==1.11.1
colorgram.py==1.2.0
pandas==1.3.5